requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.2
Django==2.0.1
django-dotenv==1.4.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
line-bot-sdk==1.5.0
psycopg2==2.7.3.2
pytz==2017.3
redis==2.10.6
requests==2.18.4
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
